+++ inte_subject/admin/autocomplete_admin.py
@@ -1,9 +1,8 @@
 from django.contrib import admin
 from edc_list_data.admin import ListModelAdminMixin
 
-from inte_lists.models import ArvRegimens
-
 from ..admin_site import inte_subject_admin
+from ..models import ArvRegimens
 
 
 @admin.register(ArvRegimens, site=inte_subject_admin)