+++ inte_subject/model_mixins/__init__.py
@@ -9,6 +9,6 @@ from .drug_refill_model_mixin import DrugRefillModelMixin
 from .drug_supply_model_mixin import DrugSupplyModelMixin
 from .glucose_model_mixin import GlucoseModelMixin
 from .medication_adherence_model_mixin import MedicationAdherenceModelMixin
-from .review_model_mixin import InitialReviewModelMixin, ReviewModelMixin
+from .review_model_mixin import ReviewModelMixin
 from .search_slug_model_mixin import SearchSlugModelMixin
 from .treatment_pay_methods import TreatmentPayMethodsModelMixin